Changes zooming, you can now use the scroll wheel
This changes the logic around zooming (and other minor things).
We no longer start you zoomed in, and better yet you can adjust the zoom yourself.  Allowing you to use the scrool wheel to zoom in and out.
This complicates the math and logic a good bit, but I think it adds a lot of value too.  I tried to test this to the best of my abilities, found TONS of bugs (some probably previously existed too).  A large part of the math complexity is around trying to keep the math accurate if someone decides to resize the main view window.
Besides the new features, I added lots of comments and javadocs, tried to make things cleaner.  I also adjusted the constants to try and get a bit better point placement in some situations.