Changes to dramatically speed up graph cleanup
We now wont tell child nodes to clean up if we are ourselves have no modifications.
This speeds it up dramatically, but puts us in more sketchy waters.  I can't think of a condition where a child part of the graph had changed when we had not changed (and would not have been called to cleanup via other mechanisms).  So it _seems_ this looping can be safely reduced.